+++ src/imports/utils/yamlutils.ts
@@ -131,4 +131,14 @@ export function getMechanicLine(document: vscode.TextDocument, lineIndex: number
 	return mechanicMap;
 }
 
+export function previousSpecialSymbol(document: vscode.TextDocument, position: vscode.Position): string {
+	const line = document.lineAt(position.line).text;
+	const text = line.substring(0, position.character);
+	const matches = text.match(/(\S)[\w\s]*$$/);
+	if (matches) {
+		return matches[1];
+	}
+	return '';
+}
+
 export { getUpstreamKey, getParentKeys, isKey, isInsideKey };